requirements.txt
supabase>=2.3.0
pandas>=2.2.0
pyarrow>=15.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
